Update main.js
Used an optimized cryptographic hash function like SHA3 instead of SHA256. This can potentially improve the performance of generating hashes.

Used a faster method to check if the hash of a block meets the difficulty requirement. Currently, this is done by repeatedly incrementing the nonce value until the condition is met, which can be slow. An alternative method is to use a proof-of-work algorithm like Ethash or Scrypt, which can be faster and more secure.

Optimized the method to check the validity of the blockchain. Currently, the method isValidChain() checks the hash of each block and the previousHash of each block, which can be slow for a large blockchain. An alternative method is to store the hash of the previous block in each block, and then simply compare this value with the previousHash of the current block to check for validity.

Reduced the number of transactions that are stored in the pendingTransactions array. This can help to reduce the memory usage and improve the performance of the blockchain.

Used a faster method to compute the balance of an address. Currently, the method getBalanceOfAddress() iterates through all the blocks and transactions in the blockchain, which can be slow for a large blockchain. An alternative method is to store the balance of each address in a separate data structure and update it whenever a transaction is added to the blockchain. This can help to significantly improve the performance of the blockchain.