Create defining the Block
for defining the Block class with a constructor method and a calculateHash() method that uses the crypto-js library to generate a SHA256 hash for the current block:
In this code, we import the SHA256 function from the crypto-js library at the top of the file using require(). We define the Block class with a constructor method that takes in the index, timestamp, data, and previousHash parameters. The calculateHash() method uses the SHA256 function to generate a hash for the current block, by concatenating the index, previousHash, timestamp, and data properties, and then passing the resulting string to SHA256(). Finally, the calculateHash() method returns the hash as a string.

With this code, you can now create new Block instances with the new keyword, passing in the required parameters:

the last two lines of code will This will create a new Block instance with an index of 0, a timestamp of the current date and time, a data property of "Hello, Kidus!", and a previousHash of "0". The calculateHash() method is automatically called when the Block instance is created, and generates a SHA256 hash for the block, which is stored in the hash property. Finally, we log the hash property to the console, which should output a unique hash value.