Create creating the Blockchain class with a constructor
we define the Blockchain class with a constructor method that initializes an empty chain array with the createGenesisBlock() method. The createGenesisBlock() method creates the first block of the chain, with an index of 0, a timestamp of the current date and time, a data property of "Genesis block", and a previousHash of "0".

The addBlock() method takes in a data parameter, and creates a new block to add to the chain. First, it gets the previousBlock by selecting the last block in the chain array. Then, it calculates the index of the new block by adding 1 to the index of the previousBlock. The timestamp is set to the current date and time. The previousHash is set to the hash property of the previousBlock. Finally, the block is created with the Block constructor, passing in the index, timestamp, data, and previousHash as parameters, and then added to the chain array with this.chain.push(block).

With  the last three lines of code , you can now create a new Blockchain instance with the new keyword, and use the addBlock() method to add new blocks to the chain:
This will create a new Blockchain instance, with the first block of the chain already created using the createGenesisBlock() method. Then, the addBlock() method is called with the string "Hello, kidus!" as the data parameter, which creates a new block with the Block constructor, and adds it to the chain array. Finally, we log the chain array to the console, which should output an array with the first and second blocks of the chain.