Create setting up a web server using the Express library
In this code, we import the Express library and the body-parser middleware for parsing JSON requests. We also import the Blockchain class from a separate file.

We create a new app instance of Express, and set the port to 3000. We then create a new blockchain instance of the Blockchain class.

We use the app.use() method to register the body-parser middleware for parsing JSON requests.

We create three endpoints for the web server:

The /blockchain endpoint is a GET request that retrieves the entire blockchain by returning the chain property of the blockchain instance as a JSON response.
The /blockchain endpoint is a POST request that adds a new block to the chain. It expects a JSON payload with a data property, which is passed to the addBlock() method of the blockchain instance. It then returns a JSON response with a message indicating that the block was added to the chain.
The /blockchain/validate endpoint is a GET request that validates the integrity of the chain by calling the validateChain() method of the blockchain instance. If the chain is valid, it returns a JSON response with a message indicating that the blockchain is valid. If the chain is not valid, it returns a 500 error with a JSON response containing an error message.
Finally, we start the web server using the app.listen() method, and log a message to the console indicating that the server is running.

the last two lines of code You can then use a web browser to access the http://localhost:3000/blockchain endpoint to retrieve the current state of the blockchain. You can use the curl command to add a new block to the chain: